CHANGE: Removed Empty columns from Images
Dropped the columns from the array containing the image data which were
all zeros. This resulted in an increase in speed when training the
machine learning model.

Change the version of the local model to keep track of how many model
versions one has developed. New model was not tested as the accuracy and
AUC indicators remained low. The model may require an in crease in
epochs to increase the accuracy of the machine learning model.