CHANGE: Modified Balance Function
Reworked the balance_data function to include a list of column names
that will be used to group the DataFrame based on the column values.
This will then be used to extract a balanced version of the dataset
based on the values using the sampling method of the DataFrame. This is
all set within a conditional statement where the default value of the
variable named columns is an empty list. In the case that the list
remains empty, it will merely shuffle the rows of the DataFrame,
otherwise it will resample and balance the dataset.

DEPRECATION: The load_data function will no longer be in use. This function is based
on the idea that one will be using folder separated images where the
folder names are the labels/classification.