CHANGE: Reworked extract_data Function
Reworked the function to take in a list which will contain all of the tag
names of the DICOM file. This will in turn allow variation where the
user can use any tags found within the DICOM file. The transformation of
the data will require standard tag name categories fed into the
transform_data function.