CHANGES: Reworked the load_training_data function
Redeveloped the function to take in any desired labels. The structure of
the function has changed significantly, in that the main input will be a
csv file containing a column parameter named pathcol to load the DICOM
files. This will be used in conjunction to the cat_labels which are a
set of column names from the CSV file itself.

Made some minor changes to the extract_data function. This function now
rescales the image before return a dictionary. It also adds the Patient
ID to the list of key:value pairs of the returned dictionary.