CHANGE: Reworked transform_data function
Recreated the function to be able to transform any dictionary data
extracted from any DICOM file. The input of the function changed from
one parameter to two, the datapoint and the definitions. The definitions
contained the DICOM tagname together with all of the possible string
values and it numerical equivalent.

NOTE:In hindsight and further analysis of this, I have noticed that this
function does not need to exist as the functionality of converting a
list of file into a pandas DataFrame will allow the conversion from
string to categorical values. A function of the same name may appear
where one does this for each column desired.