CHANGE: Fixed balance_data function
The balance_data function was collecting the number of samples and
substracting the left over from the list of samples that could not be
matched due to the categories not having enough samples. This was
corrected by simply adding the number of samples left to select to the
overall pool.