CHANGE: Developed the Data Set
Concatenated the training and test data from the calc set and the mass
set into one fully consolidated data set. Merged the consolidated data
set to the metadata V2 data set which contained the paths to the slices.
This new data set has been saved as a CSV file and will be used as the
main data set for developing the machine learning model.