CHANGE:Transformed the Dataset to Feed into Pipe
After joining the data into one consolidated and clean dataset, the full
location to the DICOM files were created using pandas DataFrames. A new
DataFrame containing the *Subject ID* and the *Series Description* columns
was created. Using a for loop, the path to the files were extracted and
the new DataFrame was merged to the old one using the two unique
identifiers. The *Series Description* was used as a unique identifier
because it distinguishes between the full mammogram image and the Region
of Interest (ROI) masks.