CHANGE: Started testing data set loading.
Finally finished developing the data set as well as changed the
rescale_image function to resize the image by a factor of two.

Further developed the test for loading the training data and am
currently working on a solution to the balance_data function (which when
tested, leads to an error where the function oversamples by 147 images).