CHANGE: Fixed Issue with balance_data Function
Redevloped the test to balance the data basedon two different columns
within the dataset. At first the test failed due to a misnomer in the
balance_data function which used the full group to calculate the
difference in sample size after selection rather than the selected
group. this has been corrected and the test now succeeds.