CHANGE: Transformed load_training_data function
Created a major change to the function where the balance_data function
is now administered based on a boolean value named *balance*. This value
defaults to True as most data will require balancing.

Finished writing the script for developing the machine learnign models,
currently encountering an issue where the python process loads all of
the data at the same time. This may be an issue with linux rather than
python itself. Will test it out in windows as well.