CHANGE: Added the img width and height to table
Developed an algorithm that extracted the image width and height and
added it to the second version of the clean data set.

Upon noticing that the data set had a variation of the image sizes,
started to work towards developing the algorithm further. Came under the
realization that some of the images were of different size than others.
This may have to be addressed further in later versions of the project
(such as padding the other sample images that can be used for training).