CHANGE: Started Testing for Training the Model
Currently encountered an issue where the image shape was not correct.
This was fixed by adding a new axis using the numpy newaxis variable.

Currently encountering an issue where the dataset loading is taking an
extensive period of time. This is causing the machine learning algorithm
to freeze up and may attributed to the size of the images (may
investigate on training the u-net using other simpler images first).

Also encountered an issue where the shape is not correct due to the
batch size (this also requires further investigation).