CHANGE: Developed algorithm to gather paths to img
Developed an algorithm using list comprehension and the os module's
listdir function to obtain a list of paths leading to the images. Then
used a set of list comprehensions to gather and filter the paths to
extract the masks and the actual images from the directory.

Developed the algorithm to use further list comprehensions to create a
series of dictionaries that can now be merged based on their image-mask
pairs.