CHANGE: Started Resizing the dataset
Developed the algorithm and created a function that would allow me to
load the images unto a pandas dataframe before loading the data in to
the TensorFlow Dataset api. Upon further experimentation, I have come to
the realization that the dataset does not contain images of the same
size. Because of this a new algorithm and parameter was added to the new
load_image function.