CHANGE: Finished Setting Up the Training of U Net
Finished the development of the algorithm to train the U Net machine
learning algorithm. Upon further analysis, I noticed that the issue
stemmed from the conversion of a pandas Series to a numpy array. This
was fixed through better organization and better use of dictionaries.

Upon further research of building the U Net, came under the realization
that the masks did not have the same dimensions as the input image.
Changed the dimensions of the mask data to be the correct dimensions
through testing and error.

A test model was successfully trained, but found multiple warnings
occurring, which may require further research. The next step is to find
the best metrics to measure the accuracy of the machine learning model
at developing a mask for the image.