CHANGE: Finished Set Up For Training U-Net
Concatenated all of the dataframes containing the images and masks.
Changed the unique identifiers to have either an m (for malignant), b
(for benign) or n (for normal) in front of the ID. Noticed that the
benign masks and image set were reversed, causing an issue with the
training. This was corrected and now the dataset is complete and set up
for training on 70% of the entire dataset.