CHANGE: Finished Conversion to pytorch
Finished converting all of the models to their pytorch version. The
biggest improvement can be found in the TumorClassifier class and the
UNet class.

The TumorClassifier is now able to take in images of any size and any
number of categorical features specified by the user of the class.

The UNet now has the capability to take images of any size in
conjunction with the ability to be as large or as small as the user of
the class desires.