LARGE CHANGE: Moved Custom Classes to New Files
Moved the Custom Datasets to the datasets file and the Trainer class to
the trainers file for better organization. the current state seems to
lead towards the development of custom dataset classes for better use of
datasets. On the other hand, custom trainers may be necessary for use in
collaboration with the custom dataset.