MINOR CHANGES: Added Parameter Type, Removed Data
Removed the Kaggle version of the CBIS dataset due to the pathing being
difficult to work with. With more time one could develop a unique ID for
the image and use that to associate with the paths, but at the moment
the idea would take too long to implement.

Added a new parameter type for the MixedDataset class. Now the csvfile
parameter takes in a polars DataFrame.