MINOR CHANGE: Reorganized MP2D Layers
Reorganized the location of the max pooling layers. After some
experimentation, noticed that the flatten layer may be difficult to use
in order to develop anything further than a fixed model. May have to
develop something similar to the double convolution with for the use of
developing machine learning models. May exclude the flattening layers
for other kinds of models that transition from the new modules.