CHANGE: Corrected Inception Blocks
Made some changes to the Stem block from the Inception AI journal. Added
a stride of two to the last convolution before the concatenated output
of the Stem block. This allowed the reduction of the image to match the
reduction from the maxpooling layer.

Made some changes to the layers for the Inception A Block from the
journal. The first branch reduced the dimensions of the image to 28x28
from 58x58, but the second, third, and fourth branches reduced the image
to 58x58, 56x56, and 54x54 respectively. The 1st branch was left as is
while the stride of the only convolutional layer for the 2nd branch
was explicitly stated to be 2 and changing the kernel size to 3 so that
the reduction was equivalent to the first branch's dimensions. The 3rd
branch only required the addition of the stride=2 to the last
convolutional layer. The 4th layer required the most changes, with the
addition of the stride=2 parameter to the last convolutional layer and a
change in kernel size from 3 to 1.

Made some changes to the layers of the Reduction A block from the
journal. Changed the padding on the AVG Pooling layer from 1 to zero for
the 1st branch. This caused the 1st branch to output 13x13 images rather
than 14x14. Left the 2nd branch as is because it had the ideal
output which matched the new output of the 1st branch. The 3rd branch
had a change in kernel size for the second convolutional layer

Made some changes to the layers of the Inception B block from the
journal. Changed the stride of the 1st branch to 1 and added padding of
1 as well to make the images 13x13. The 2nd branch was left alone. The
3rd branch had padding added to it in the last convolutional layer. The
padding was calculated to be (0,7-1). the 4th branch had padding added
to the last convolutional layers. The padding was (0,7-1) and (7-1,0).

Made some changes to the layers of the Reduction B block from the
journal. Within the forward pass function, added the subscript of 3 to
the inputs of some convolutions which were missing due to a previous
error. The 1st branch had the padding of the maxpooling layer changed
from 1 to 0. This reduced the image dimensions to 6x6 from 7x7. The 2nd
branch did not require any changes as it already had the appropriate
dimensions. The 3rd branch required the addition of padding.
Respectively a (0,3) and (3,0) padding were added to the middle 2 layers
of the 3rd branch. This was calculated from the following (1-1/2, 7-1/2)
and (7-1/2, 1-1/2).

Made some changes to the Inception C block from the journal. All
branches of the block were changed to have images that fit the dimension
3x3. The 1st branch had the padding of the avgpooling layer changed
from 1 to 0. The 2nd branch had a stride=2 added to its convolutional
layer. The 3rd branch had a stride of 2 added to its left minor branch
and right minor branch. The minor branches received a padding of (0,1)
and (1,0) respectively. The 4th branch also had a stride of 2 added to
the minor left branch and the minor right branch. Each minor branch also
had padding added to it. Respectively, the padding was (2,1) and (1,2).

NEW: InceptionV4 Module

Created a new module that simulates the Inception V4 model similar to
the one mentioned in the journal. Added a set of linear layers that
reduce the number of features down to the classes. Added the set of
layers to the forward pass and corrected the numbering of the linear
layers. Set the dim parameter to 1 for the softmax layer.