MAJOR CHANGE: Modified extract_key_data function
Changed the function extract_key_data to extract_metadata. Now the
function extracts the path to the image along with provide a unique id.
A parameter for extracting key metadata within the dicom file is also
added.

Created a function to extract the file paths from a root directory as
well. Added the capability for one to also save the list of paths within
a text file.