NEW: Pendulum Environment
Created the basic form of the pendulum environment that is based on the
tutorial of how to build custom environments in torchrl. This
environment works based on the tensordict functionality and setting up
parameters for setting how the neural network interacts with the
environment. This also helps set limitations on the neural network
itself, and allows a reward/cost system.