MAJOR CHANGES: Finished First Version of Pipeline
Finished Creating the data pipeline for the CBIS-DDSM dataset containing
dicom files as well as metadata from different csv files. New functions
were created to extract and merge the metadata with the paths leading
to the dicom files. Some functions were modified to assure proper
reading of certain parameters (explained in more detail below).

CHANGE: extract_paths method

Encapsulated the second for loop within an if statement that affirms
that the string contains 'image' within it. This prevents any error from
any other path taken from the root directory.

CHANGE: create_unique_id method

Had an error in the process of creating a new column from existing
columns. In the case that the resulting column is a string, one is to
use the polars function `concat_str` which allows one to concatenate all
of the columns into a singular string.

NEW: extract_description_sets Method

Created a new method that looks through the root directory for any csv
file containing the word "description" and places the path to the file
within a list.

NEW: concat_description_sets Method

Created a new method that loads the files extracted using the
`extract_description_sets` method and to concatenate the files together.
This returns a polars DataFrame.