CHANGE: Removed Static Methods
Removed the static methods from the classifier pipeline. The classifier
class had a set of functions which were all staticmethods in nature.
This was not necessary as the user can just use the attributes
associated with the class instead of creating a static method that would
require the user to input parameters manually as opposed to
automatically