MAJOR CHANGE: Trainer Abstract Class
Made a major change to the Trainer Abstract class by including the
boolean dictating the use of the gpu at the initiation of the class.
This way the data and the model can be sent to the device throughout
different functions.

Also built the base template for the ROI Trainer.