CHANGE: Modified Methods to Exclude CSV Files
Added a condition to the base class which excludes the csv files from
the list. Changed the methods from the subclasses to no longer exclude
the component related to the creation of a newline.

Created an algorithm to observe the image dimensions before feeding into
the machine learning model for training. Through this algorithm I
discovered that the roi images are not organized (meaning that the mask
and roi image are not consistently labelled 2 and 1 respectively). Added
a new condition to the create_unique_id method for the ROI pipeline to
include the width and height of each image. In the start method of the
ROI pipeline, these new columns function to assure that the full images
are matched to the roi by selecting only the cases where the pixel width
is different. The opposite is true for matching the full image to the
mask image.