FIX: Unet model Down Convolutions
Corrected the issue where the down convolutions did not match the same
pattern as the up convolutions, causing a situation where the expected
input is not the same as the actual input. This was a case that only
applied during the situation where the input channels is equivalent to
the output channels (mainly due to a personal confusion on the loop
implementation).